HowHot-backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
pillow-heif
python-multipart
boto3
asyncpg
sqlalchemy
python-dotenv
psycopg2-binary